pyproject.toml
[tool.poetry]
name = "drf-atomic-habits"
version = "0.1.0"
description = ""
authors = ["IvanPro91 <ivan-don4encko@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.5"
celery = "^5.5.3"
redis = "^6.4.0"
django-celery-beat = "^2.8.1"
python-telegram-bot = "^22.3"
djangorestframework = "^3.16.1"
djangorestframework-simplejwt = "^5.5.1"
django-cors-headers = "^4.7.0"
python-dotenv = "^1.1.1"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
isort = "^6.0.1"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '.git'

[tool.isort]
line_length = 119
